Collapse middleware into message
